refactor - Cars 객체 생성시 cars필드의 데이터를 만들때 for문대신 이터러블 메서드 forEach사용하여 구현 변경
